+++ src/pages/Issue/Details/components/ErrorStack.tsx
@@ -1,6 +1,6 @@
 import React, { FC } from 'react';
 import { ErrorContext, StackTrace } from '@store/type';
-import EventDetailHeader from '../EventDetailHeader';
+import EventDetailHeader from './EventDetailHeader';
 import StackHeader from './StackHeader';
 import StackContext from './StackContext';
 
@@ -9,14 +9,14 @@ interface Props {
   errorContexts: ErrorContext[];
 }
 
-
 const ErrorStack: FC<Props> = ({ stacktrace, errorContexts }: Props) => (
   <div>
     <EventDetailHeader title="ERROR STACK" />
     {stacktrace.length !== 0 &&
       stacktrace.map((stack, idx) => (
         <StackHeader stacktrace={stack} key={idx} idx={idx}>
-          {errorContexts.length !== 0 && errorContexts[idx].errorContext.length ? (
+          {errorContexts.length !== 0 &&
+          errorContexts[idx].errorContext.length ? (
             <StackContext
               contexts={errorContexts[idx]}
               lineno={stack.lineno ? stack.lineno : 0}
