+++ src/pages/Issue/Details/components/EventDetailHeader.tsx
