+++ src/pages/Issue/Details/components/EventHeader.tsx
