+++ src/pages/Issue/Details/components/StackHeader.tsx
