+++ src/pages/Issue/components/IssueTabs.tsx
@@ -3,9 +3,9 @@ import { useHistory, useRouteMatch, Route } from 'react-router-dom';
 import { makeStyles } from '@material-ui/core/styles';
 import { Paper, Tabs, Tab } from '@material-ui/core';
 import { Issue } from '@store/type';
-import DetailsTabContainer from './Details/DetailsTabContainer';
-import EventsTabContainer from './Events/EventsTabContainer';
-import IssueTagsContainer from './Tags/IssueTagsContainer';
+import DetailsTabContainer from '../Details/DetailsTabContainer';
+import EventsTabContainer from '../Events/EventsTabContainer';
+import IssueTagsContainer from '../Tags/IssueTagsContainer';
 
 const useStyles = makeStyles({
   root: {
